requirements.txt
flask
tensorflow
opencv-python
numpy
pandas
matplotlib
scikit-learn
pillow
requests